Change long to integer for startedAt.

+++ conf/evolutions/default/1.sql
@@ -19,7 +19,7 @@ CREATE TABLE task (
     id serial PRIMARY KEY,
     userId integer NOT NULL,
     serverId integer NOT NULL,
-    startedAt long NOT NULL,
+    startedAt integer NOT NULL,
     duration integer NOT NULL default 0,
     failed boolean NOT NULL default false
 );
